completing dependency upgrade after zio release for scala native 0.5
