+++ project/BuildHelper.scala
@@ -315,24 +315,25 @@ object BuildHelper {
 
   object Dependencies {
     val catsEffect             = "3.4.9"
-    val circe                  = "0.14.3"
+    val circe                  = "0.14.9"
+    val circeGenericExtras     = "0.14.4"
     val http4s                 = "0.23.26"
     val jawnAST                = "1.5.1"
-    val jsoniterScala          = "2.23.3"
+    val jsoniterScala          = "2.30.7"
     val kindProjector          = "0.13.3"
     val magnolia2              = "1.1.10"
     val magnolia3              = "1.3.7"
     val playJson               = "2.9.4"
     val playJsonExtensions     = "0.43.1"
     val playJsonExtensions2_12 = "0.42.0"
-    val refined                = "0.10.2"
+    val refined                = "0.11.2"
     val scalaCollectionCompat  = "2.12.0"
-    val scalaJavaTime          = "2.5.0"
+    val scalaJavaTime          = "2.6.0"
     val scalaMacros            = "2.1.1"
-    val scalaz                 = "7.3.7"
+    val scalaz                 = "7.2.36"
     val silencer               = "1.7.16"
     val snakeYaml              = "2.2"
-    val zio                    = "2.1.0"
+    val zio                    = "2.1.7"
     val zioInteropCats         = "23.1.0.2"
     val zioReflect             = "1.0.0-M10"
   }
